requirements.txt
fastapi[all]
streamlit
spire.doc
mistralai